requirement.txt
# Pythonライブラリ
flask
flask_bcrypt
mysql-connector-python
google-api-core
google-api-python-client
google-auth

# MySQL
https://prog-8.com/docs/mysql-env-win を見ながらインストール

# MySQLにcmdで接続
mysql --user=root --password

# データベース作成
CREATE DATABASE userdb;
CREATE USER 'g09'@'localhost' IDENTIFIED BY '1234';
GRANT ALL PRIVILEGES ON userdb.* TO 'g09'@'localhost';
USE userdb;
CREATE TABLE users (
    id INT AUTO_INCREMENT PRIMARY KEY,
    username VARCHAR(80) NOT NULL,
    email VARCHAR(120) NOT NULL,
    password VARCHAR(60) NOT NULL
    );

CREATE TABLE jobs(
    job_name VARCHAR(255) NULL,
    hourly_wage INT NOT NULL
    );

alter table jobs convert to character set utf8;

# 実行はcmdで
net start mysql57   # 管理者権限で
python main.py